Rework SimpleLog > Log; Reintegrate worklog edit;

+++ database2/src/main/java/lt/markmerkk/WorklogStorage.kt
@@ -23,7 +23,7 @@ class WorklogStorage(
         }
     }
 
-    fun insertOrUpdateSync(log: Log): Int {
+    fun insertOrUpdateSync(log: Log): Long {
         val insertRemoteLog = log.isRemote
         val insertLocalLog = !log.isRemote
         val localId = log.id
@@ -59,17 +59,17 @@ class WorklogStorage(
                 )
             }
         }
-        return Const.NO_ID.toInt()
+        return Const.NO_ID
     }
 
-    fun insertOrUpdate(log: Log): Single<Int> {
+    fun insertOrUpdate(log: Log): Single<Long> {
         return Single.defer {
             Single.just(insertOrUpdateSync(log))
         }
     }
 
     // todo fix this as this needs tinkering, as whenever update log, remote logs should be marked for deletion
-    fun updateSync(log: Log): Int {
+    fun updateSync(log: Log): Long {
         val updateRemoteLog = log.isRemote
         val updateLocalLog = !log.isRemote
         val localId = log.id
@@ -78,7 +78,7 @@ class WorklogStorage(
         val existAsRemote = dbInteractor.existAsRemote(remoteId = remoteId)
 
         if (!existAsLocal && !existAsRemote) {
-            return Const.NO_ID.toInt()
+            return Const.NO_ID
         }
 
         // insert local log
@@ -110,10 +110,10 @@ class WorklogStorage(
                 )
             }
         }
-        return Const.NO_ID.toInt()
+        return Const.NO_ID
     }
 
-    fun update(log: Log): Single<Int> {
+    fun update(log: Log): Single<Long> {
         return Single.defer {
             Single.just(updateSync(log))
         }
@@ -133,7 +133,7 @@ class WorklogStorage(
         return dbInteractor.findByRemoteId(remoteId)
     }
 
-    fun deleteSync(log: Log): Int {
+    fun deleteSync(log: Log): Long {
         val localId = log.id
         val remoteId = log.remoteData?.remoteId ?: Const.NO_ID
         val existAsLocal = dbInteractor.existAsLocal(localId = localId)
@@ -144,20 +144,20 @@ class WorklogStorage(
         if (existAsRemote) {
             return dbInteractor.update(log.markAsDeleted())
         }
-        return Const.NO_ID.toInt()
+        return Const.NO_ID
     }
 
-    fun delete(log: Log): Single<Int> {
+    fun delete(log: Log): Single<Long> {
         return Single.defer {
             Single.just(deleteSync(log))
         }
     }
 
-    fun hardDeleteRemoteSync(remoteId: Long): Int {
+    fun hardDeleteRemoteSync(remoteId: Long): Long {
         return dbInteractor.deleteByRemoteId(remoteId)
     }
 
-    fun hardDeleteRemote(remoteId: Long): Single<Int> {
+    fun hardDeleteRemote(remoteId: Long): Single<Long> {
         return Single.defer {
             Single.just(hardDeleteRemoteSync(remoteId))
         }
